Rename classloading module to class-loader
