+++ instrumentation/class-loader/jboss-testing/jboss-testing.gradle
@@ -1,7 +1,7 @@
 apply from: "${rootDir}/gradle/java.gradle"
 
 dependencies {
-  testCompile project(':instrumentation:classloading')
+  testCompile project(':instrumentation:class-loader')
 
   testCompile group: 'org.jboss.modules', name: 'jboss-modules', version: '1.3.10.Final'
 }
