+++ instrumentation/class-loader/jboss-testing/src/test/groovy/JBossClassloadingTest.groovy
