+++ instrumentation/class-loader/osgi-testing/src/test/groovy/OSGIClassloadingTest.groovy
