+++ settings.gradle
@@ -51,10 +51,10 @@ include ':instrumentation:apache-httpclient-4.0'
 include ':instrumentation:aws-java-sdk-1.11'
 include ':instrumentation:aws-java-sdk-2.2'
 include ':instrumentation:cdi-1.2'
-include ':instrumentation:classloading'
-include ':instrumentation:classloading:jboss-testing'
-include ':instrumentation:classloading:osgi-testing'
-include ':instrumentation:classloading:tomcat-testing'
+include ':instrumentation:class-loader'
+include ':instrumentation:class-loader:jboss-testing'
+include ':instrumentation:class-loader:osgi-testing'
+include ':instrumentation:class-loader:tomcat-testing'
 include ':instrumentation:couchbase-2.0'
 include ':instrumentation:couchbase-2.6'
 include ':instrumentation:datastax-cassandra-3.0'
